AI_System/langchain/requirements.txt
flask
numpy
pandas
scikit-learn
python-dotenv
langchain[llms]
openai
langchain-community
langchain-openai
langchain-core
langchain-cli
langchain-text-splitters
langchain-openai
beautifulsoup4
pandas
langchainhub
typing